GetInsuranceEvidenceInitialEmailUseCaseTest
